backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.4.0
httpx>=0.27.0
openai>=1.6.0
supabase>=2.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
bcrypt>=4.1.2
pandas>=2.1.4
numpy>=1.26.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
annoy>=1.17.3
scikit-learn>=1.3.2
weasyprint>=62.0
jinja2>=3.1.2
aiofiles>=23.2.1
python-multipart>=0.0.7
orjson>=3.9.10

